Modified comments, changed class and method names, and used different aliases for loaded functions

+++ dataOps/dataOps.py
@@ -1,9 +1,8 @@
 '''
-    File name: dataOperators.py
+    File name: dataOps.py
     Author: James Chryssanthacopoulos
     Date created: 01/03/2017
     Date last modified: 01/04/2017
-    Python Version: 2.7
 '''
 
 
@@ -12,7 +11,9 @@ import string
 import os
 import random
 import shutil
-from utils import pjoin
+import filecmp
+from os.path import join as pjoin
+from os import listdir as ldir
 
 
 # Superclass
@@ -32,7 +33,8 @@ class ControlledFileSystem(object):
     
     def __init__(self, directory, targetFileSize = 0, targetDirSize = 0):
         """
-        Class constructor: ControlledFileSystem(directory, targetFileSize, targetDirSize)
+        Class constructor:
+            ControlledFileSystem(directory, targetFileSize, targetDirSize)
         """
         self.directory = directory
         self.targetFileSize = targetFileSize
@@ -87,11 +89,14 @@ class ControlledFileSystem(object):
 
 
     def _dirSize(self):
-        dirListing = os.listdir(self.directory)
+        """
+        This method returns the size of the directory in bytes
+        """
+        dirListing = ldir(self.directory)
         
         dirSize = 0
         for item in dirListing:
-            itemFullPath = os.path.join(self.directory, item)
+            itemFullPath = pjoin(self.directory, item)
             
             if os.path.isfile(itemFullPath):
                 dirSize += os.path.getsize(itemFullPath)
@@ -100,7 +105,10 @@ class ControlledFileSystem(object):
     
     
     def _numFiles(self):
-        return len(os.listdir(self.directory))
+        """
+        This method returns the number of files in the directory
+        """
+        return len(ldir(self.directory))
 
 
     def _fileFull(self, fileSize):
@@ -112,42 +120,43 @@ class ControlledFileSystem(object):
     
     def _dirFull(self):
         """
-        This method checks if the size of the directory is within 1% of the target size or more
+        This method checks if the size of the directory is within 1% of the target size or more or if the max file number has been met
         """
         return self._dirSize() >= 0.99 * self.targetDirSize or self._numFiles() >= self.maxFiles
 
 
 
 # This class generates a random master dataset as a set of files on disk
-class FileInitializer(ControlledFileSystem):
+class DataInitializer(ControlledFileSystem):
     def __init__(self, directory, targetFileSize, targetDirSize):
-        super(FileInitializer, self).__init__(directory, targetFileSize, targetDirSize)
+        super(DataInitializer, self).__init__(directory, targetFileSize, targetDirSize)
         """
-        Class constructor
+        Class constructor:
+            DataInitializer(directory, targetFileSize, targetDirSize)
         """        
         # Make directory if it doesn't already exist
         if not os.path.exists(directory):
             os.makedirs(directory)
         else:
             # Remove any preexisting files
-            filelist = os.listdir(directory)
+            filelist = ldir(directory)
             for filename in filelist:
-                os.remove(os.path.join(directory, filename))
+                os.remove(pjoin(directory, filename))
     
     
-    def writeFiles(self):
+    def writeData(self):
         """
         This method will create all the files in the directory according to the specifications
         """
         
-        print 'Writing files for directory %s . . .' % self.directory
+        print 'Writing data for directory %s . . .' % self.directory
         
         # Keep creating files until directory size has reached the target
         # or if maximum number of files has been reached
         
         while not self._dirFull():
             # Get new filename of file to write
-            filename = os.path.join(self.directory, 'file%02d.txt' % (self._numFiles() + 1))
+            filename = pjoin(self.directory, 'file%02d.txt' % (self._numFiles() + 1))
             
             # Write file
             self.writeFile(filename)
@@ -155,31 +164,35 @@ class FileInitializer(ControlledFileSystem):
 
 
 # This class updates a master dataset
-class FileUpdater(ControlledFileSystem):    
+class DataUpdater(ControlledFileSystem):    
     def __init__(self, directory, targetFileSize, targetDirDelta):
-        super(FileUpdater, self).__init__(directory, targetFileSize)
+        """
+        Class constructor:
+            DataUpdater(directory, targetFileSize, targetDirDelta)
+        """
+        super(DataUpdater, self).__init__(directory, targetFileSize)
         
         # Set dtarget directory size to current size plus delta
         self.targetDirSize = self._dirSize() + targetDirDelta
 
 
-    def updateFiles(self):
+    def updateData(self):
         """
         This method will update all the files in the directory according to the specifications
         """
         
-        print 'Updating files for directory %s . . .' % self.directory
+        print 'Updating data for directory %s . . .' % self.directory
         
         # Keep creating files until directory size has reached the target
         # or if maximum number of files has been reached
         
         # Update last file if need be
-        lastCreatedFile = os.path.join(self.directory, os.listdir(self.directory)[-1])
+        lastCreatedFile = pjoin(self.directory, ldir(self.directory)[-1])
         self.writeFile(lastCreatedFile)
         
         while not self._dirFull():
             # Get new filename of file to write
-            filename = os.path.join(self.directory, 'file%02d.txt' % (self._numFiles() + 1))
+            filename = pjoin(self.directory, 'file%02d.txt' % (self._numFiles() + 1))
             
             # Write file
             self.writeFile(filename)
@@ -187,35 +200,56 @@ class FileUpdater(ControlledFileSystem):
 
 
 # This class backups a master dataset
-class FileBackup(object):
+class DataBackup(object):
+    """
+    DataBackup()
+    """
     
     def backup(self, dataDirectory, backupDirectory):
         """
         This method backups a master dataset into a backup directory
         """
         
+        print 'Backing up directory %s into %s . . .' % (dataDirectory, backupDirectory)
+        
         # Name of directory
         dataName = os.path.basename(dataDirectory)
         
+        # Name in backup directory
         backupName = pjoin(backupDirectory, dataName)
         
+        # Create directory in backup if it doesn't exist
         if not os.path.exists(backupName):
-            # No previous backup
             os.makedirs(backupName)
         
+        # List of subdirectories
+        srcDirNames = ldir(dataDirectory)
+        
+        # Remove stale subdirectories in backup
+        for d in [drct for drct in ldir(backupName) if drct not in srcDirNames]:
+            shutil.rmtree(pjoin(backupName, d))
+        
         # Iterate through subdirectories
-        for dirName in os.listdir(dataDirectory):
+        for dirName in srcDirNames:
             srcDirName = pjoin(dataDirectory, dirName)
             destDirName = pjoin(backupName, dirName)
             
+            # Create subdirectory if it doesn't exist
             if not os.path.exists(destDirName):
-                # Make new subdirectory
                 os.makedirs(destDirName)
             
+            # List of files
+            srcFiles = ldir(srcDirName)
+            
+            # Remove stale files in backup
+            for f in [fname for fname in ldir(destDirName) if fname not in srcFiles]:
+                os.remove(pjoin(destDirName, f))
+            
             # Iterate through files
-            for fileName in os.listdir(srcDirName):
+            for fileName in srcFiles:
                 srcFileName = pjoin(srcDirName, fileName)
                 destFileName = pjoin(destDirName, fileName)
                 
-                # Copy file
-                shutil.copyfile(srcFileName, destFileName)
+                # Copy file if it doesn't exist in backup or a different file of the same name resides there
+                if not os.path.exists(destFileName) or not filecmp.cmp(srcFileName, destFileName):
+                    shutil.copyfile(srcFileName, destFileName)
